system/django/kisho/cinema/urls.py
from django.urls import path
from rest_framework import routers
from . import views

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'movies', views.MovieViewSet)

urlpatterns = [
    path('', views.index, name='index'),
]