system/django/requirements.txt
asgiref==3.2.10
asn1crypto==0.24.0
click==7.1.2
cryptography==2.1.4
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.6
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
pycrypto==2.6.1
pygobject==3.26.1
pytz==2020.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.3.1
yapf==0.30.0
zipp==3.1.0
